+++ MoltenVK/MoltenVK/GPUObjects/MVKPipeline.mm
@@ -856,20 +856,26 @@ MVKGraphicsPipeline::MVKGraphicsPipeline(MVKDevice* device,
 	_staticStateData.primitiveType = mvkMTLPrimitiveTypeFromVkPrimitiveTopology(_vkPrimitiveTopology);
 	_staticStateData.enable.set(MVKRenderStateEnableFlag::PrimitiveRestart, primitiveRestart);
 
-	// In Metal, primitive restart cannot be disabled, so issue a warning if the app
-	// has disabled it statically, or indicates that it might do so dynamically.
-	// Just issue a warning here, as it is very likely the app is not actually
-	// expecting to use primitive restart at all, and is disabling it "just-in-case".
-	// As such, forcing an error here would be unexpected to the app (including CTS).
-	// BTW, although Metal docs avoid mentioning it, testing shows that Metal does not support primitive
-	// restart for list topologies, meaning VK_EXT_primitive_topology_list_restart cannot be supported.
-	if (( !primitiveRestart || _dynamicStateFlags.has(MVKRenderStateFlag::PrimitiveRestartEnable)) &&
-		 (_vkPrimitiveTopology == VK_PRIMITIVE_TOPOLOGY_LINE_STRIP ||
-		  _vkPrimitiveTopology == VK_PRIMITIVE_TOPOLOGY_TRIANGLE_STRIP ||
-		  _vkPrimitiveTopology == VK_PRIMITIVE_TOPOLOGY_TRIANGLE_FAN ||
-		  _dynamicStateFlags.has(MVKRenderStateFlag::PrimitiveTopology))) {
-		reportWarning(VK_ERROR_FEATURE_NOT_PRESENT, "Metal does not support disabling primitive restart.");
+#if MVK_USE_METAL_PRIVATE_API
+	if (!getMVKConfig().useMetalPrivateAPI) {
+#endif
+		// In Metal, primitive restart cannot be disabled, so issue a warning if the app
+		// has disabled it statically, or indicates that it might do so dynamically.
+		// Just issue a warning here, as it is very likely the app is not actually
+		// expecting to use primitive restart at all, and is disabling it "just-in-case".
+		// As such, forcing an error here would be unexpected to the app (including CTS).
+		// BTW, although Metal docs avoid mentioning it, testing shows that Metal does not support primitive
+		// restart for list topologies, meaning VK_EXT_primitive_topology_list_restart cannot be supported.
+		if ((!primitiveRestart || _dynamicStateFlags.has(MVKRenderStateFlag::PrimitiveRestartEnable)) &&
+			 (_vkPrimitiveTopology == VK_PRIMITIVE_TOPOLOGY_LINE_STRIP ||
+			  _vkPrimitiveTopology == VK_PRIMITIVE_TOPOLOGY_TRIANGLE_STRIP ||
+			  _vkPrimitiveTopology == VK_PRIMITIVE_TOPOLOGY_TRIANGLE_FAN ||
+			  _dynamicStateFlags.has(MVKRenderStateFlag::PrimitiveTopology))) {
+			reportWarning(VK_ERROR_FEATURE_NOT_PRESENT, "Metal does not support disabling primitive restart.");
+		}
+#if MVK_USE_METAL_PRIVATE_API
 	}
+#endif
 
 	// Must run after _isRasterizing and _dynamicState are populated
 	initSampleLocations(pCreateInfo);