+++ Docs/Whats_New.md
@@ -31,6 +31,7 @@ Released TBD
 - Added working demo target for VisionOS in Cube Demo.
 - Raise minimum target to _macOS 11.0 / iOS 14.0 / tvOS 14.0_
 - Add support for `VK_IMAGE_CREATE_BLOCK_TEXEL_VIEW_COMPATIBLE_BIT`.
+- Add support for disabling primitive restart when `MVK_USE_METAL_PRIVATE_API` is enabled.
 - Reject image usages whose format features are not supported.
 - Return `VK_ERROR_OUT_OF_POOL_MEMORY` when running out of descriptor sets.
 - Make vkGetPhysicalDeviceToolProperties() an instance function.